"Mouse keys"-like behavior
Instead of trying to move the mouse to where the user is looking, use
the direction of off screen gazes to move the mouse. This effectively
creates eight gaze regions similar to the numpad buttons used for mouse
keys.

The speed of the mouse is proportional to how far off screen the user is
looking and when the mouse goes off the edge of the screen it reappears
on the opposite edge.